refactor: extract SnapSystem from AppController (Phase 2-A)
Move _filterNearbySnapTargets, _findNearestSnapCandidate, _pickBestSnapTarget,
and _projectToScreen into pure exported functions in src/controller/snap/SnapSystem.js.
All functions now take camera as an explicit parameter, making them stateless
and testable in isolation. MeasurePlacementHandler imports pickBestSnapTarget
directly; AppController call sites updated to pass this._camera.

AppController: 3972 → 3857 lines (−115 lines)
